Move get_available_port from cli_test to utils.hpp
