+++ tests/common/utils.hpp
@@ -26,6 +26,19 @@
 #include <fc/thread/thread.hpp>
 #include <boost/test/included/unit_test.hpp>
 
+#ifdef _WIN32
+   #ifndef _WIN32_WINNT
+      #define _WIN32_WINNT 0x0501
+   #endif
+   #include <winsock2.h>
+   #include <ws2tcpip.h>
+#else
+   #include <sys/types.h>
+   #include <sys/socket.h>
+   #include <netinet/in.h>
+   #include <netinet/ip.h>
+#endif
+
 namespace fc {
 
    /** Waits for F() to return true before max_duration has passed.
@@ -38,4 +51,34 @@ namespace fc {
          fc::usleep(fc::milliseconds(100));
       BOOST_REQUIRE( f() );
    }
-}
+
+namespace network {
+   //////
+   /// @brief attempt to find an available port on localhost
+   /// @returns an available port number, or -1 on error
+   /////
+   int get_available_port()
+   {
+      struct sockaddr_in sin;
+      int socket_fd = socket(AF_INET, SOCK_STREAM, 0);
+      if (socket_fd == -1)
+         return -1;
+      sin.sin_family = AF_INET;
+      sin.sin_port = 0;
+      sin.sin_addr.s_addr = htonl(INADDR_LOOPBACK);
+      if (::bind(socket_fd, (struct sockaddr*)&sin, sizeof(struct sockaddr_in)) == -1)
+         return -1;
+      socklen_t len = sizeof(sin);
+      if (getsockname(socket_fd, (struct sockaddr *)&sin, &len) == -1)
+         return -1;
+   #ifdef _WIN32
+      closesocket(socket_fd);
+   #else
+      close(socket_fd);
+   #endif
+      return ntohs(sin.sin_port);
+   }
+
+} // namespace fc::network
+
+} // namespace fc
